# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/category": {
      "post": {
        "operationId": "createCategory",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "category"
        ]
      },
      "get": {
        "operationId": "findCategories",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "category"
        ]
      }
    },
    "/category/{id}": {
      "patch": {
        "operationId": "findCategory",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "category"
        ]
      },
      "delete": {
        "operationId": "deleteCategory",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "category"
        ]
      }
    },
    "/courses": {
      "post": {
        "operationId": "createCourse",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "get": {
        "operationId": "getCourses",
        "summary": "",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseWithCategoryTeacher"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "operationId": "getCourseById",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CourseWithCategoryTeacher"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "patch": {
        "operationId": "updateCourse",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCourseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "operationId": "deleteCourse",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "courses"
        ]
      }
    },
...
    "/videos": {
      "post": {
        "operationId": "createVideo",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      },
      "get": {
        "operationId": "findVideos",
        "summary": "",
        "parameters": [
          {
            "name": "courseId",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}": {
      "get": {
        "operationId": "findOneVideo",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Video"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      },
      "patch": {
        "operationId": "updateVideo",
        "summary": "",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVideoDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      },
      "delete": {
        "operationId": "VideosController_deleteVideo",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Video"
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
    "/videos/{id}/students": {
      "post": {
        "operationId": "VideosController_getStudentsByVideoId",
        "summary": "讀取所有看過特定一支影片的學生",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "videos"
        ]
      }
    },
...
    "/favorites": {
      "post": {
        "operationId": "addToFavorite",
        "summary": "Add favorite with video id and user id.",
        "description": "Add a video to favorite list for a user",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFavoriteDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "favorites"
        ]
      },
      "get": {
        "operationId": "getFavorite",
        "summary": "Retrieve favorites",
        "parameters": [
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "nullable": true,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "favorites"
        ]
      }
    },
    "/favorites/{id}": {
      "delete": {
        "operationId": "FavoritesController_delete",
        "summary": "Remove favorite with id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Favorite"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "favorites"
        ]
      }
    },
...
  },
  "info": {
    "title": "Catty Course Api",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "Category": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "categoryId",
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "CreateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "teacherId": {
            "type": "number"
          },
          "categoryId": {
            "type": "number"
          },
          "thumbnail": {
            "type": "number"
          }
        },
        "required": [
          "name",
          "description",
          "teacherId",
          "categoryId",
          "thumbnail"
        ]
      },
      "Course": {
        "type": "object",
        "properties": {
          "courseId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "thumbnail": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "teacherId": {
            "type": "number"
          },
          "categoryId": {
            "type": "number"
          }
        },
        "required": [
          "courseId",
          "name",
          "description",
          "thumbnail",
          "createdAt",
          "teacherId",
          "categoryId"
        ]
      },
      "CourseTeacher": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name",
          "email"
        ]
      },
      "CourseCategory": {
        "type": "object",
        "properties": {}
      },
      "CourseWithCategoryTeacher": {
        "type": "object",
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "courseId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "teacher": {
            "$ref": "#/components/schemas/CourseTeacher"
          },
          "category": {
            "$ref": "#/components/schemas/CourseCategory"
          },
          "thumbnail": {
            "type": "string"
          }
        },
        "required": [
          "createdAt",
          "courseId",
          "name",
          "description",
          "teacher",
          "category",
          "thumbnail"
        ]
      },
      "UpdateCourseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "teacherId": {
            "type": "number"
          },
          "categoryId": {
            "type": "number"
          },
          "thumbnail": {
            "type": "number"
          }
        }
      },
...
      "CreateVideoDto": {
        "type": "object",
        "properties": {
          "fileLink": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "length": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "courseId": {
            "type": "number"
          }
        },
        "required": [
          "fileLink",
          "name",
          "length",
          "description",
          "courseId"
        ]
      },
      "Video": {
        "type": "object",
        "properties": {
          "videoId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "fileLink": {
            "type": "string"
          },
          "length": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "courseId": {
            "type": "number"
          }
        },
        "required": [
          "videoId",
          "name",
          "fileLink",
          "length",
          "description",
          "createdAt",
          "courseId"
        ]
      },
      "UpdateVideoDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "length": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "fileLink": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "length",
          "description",
          "fileLink"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "role": {
            "type": "string",
            "enum": [
              "guest",
              "student",
              "teacher"
            ]
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "avatar": {
            "type": "string"
          }
        },
        "required": [
          "userId",
          "role",
          "name",
          "email",
          "createdAt",
          "password",
          "avatar"
        ]
      },
...
      "CreateFavoriteDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "videoId": {
            "type": "number"
          }
        },
        "required": [
          "userId",
          "videoId"
        ]
      },
      "Favorite": {
        "type": "object",
        "properties": {
          "favoriteId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "userId": {
            "type": "number"
          },
          "videoId": {
            "type": "number"
          }
        },
        "required": [
          "favoriteId",
          "createdAt",
          "userId",
          "videoId"
        ]
      },
...
    }
  }
}